+++ Runtime/Scripts/Material/UniveralRPMaterialGenerator.cs
@@ -36,7 +36,7 @@ namespace GLTFast.Materials {
         
 #if USING_URP_12_OR_NEWER
         protected override string GetMetallicShaderName(MetallicShaderFeatures metallicShaderFeatures) {
-            return "Shader Graphs/glTF-generic";
+            return "Shader Graphs/glTF-pbrMetallicRoughness";
         }
         
         protected override void SetDoubleSided(Schema.Material gltfMaterial, Material material) {